requirements.txt
blake3
tqdm
rich
sqlite-utils
numpy
Pillow
onnxruntime
nvidia-ml-py
opencv-python-headless
fastapi
uvicorn
guessit
rapidfuzz
tmdbsimple
IMDbPY
requests
